% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.labelled}
\alias{as.data.frame.labelled}
\title{as.data.frame method for labelled vectors}
\usage{
\method{as.data.frame}{labelled}(x, ...)
}
\arguments{
\item{x}{a labelled vector}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
as.data.frame method for labelled vectors
}

