% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_labelled.R
\name{as_labelled}
\alias{as_labelled}
\alias{as_labelled.factor}
\title{Convert factor to labelled}
\usage{
as_labelled(x)

\method{as_labelled}{factor}(x)
}
\arguments{
\item{x}{factor to convert}
}
\description{
Convert a factor to a numerical
}
\note{
Note that due to the way that \R store internally factors,
\code{as_labelled(as_factor(x))} will usually not equal
to \code{x} (see examples).
}
\examples{
x <- factor(c("a", "a", "c", "b", "b"))
as_labelled(x)

y <- labelled(c(2, 1, 9, 1, 2), c(yes = 1, no = 2, dk = 9))
y
as_labelled(as_factor(y))
}

