% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_unused_value_labels.R
\name{drop_unused_value_labels}
\alias{drop_unused_value_labels}
\title{Drop unused value labels}
\usage{
drop_unused_value_labels(x)
}
\arguments{
\item{x}{A vector or a data frame.}
}
\description{
Drop value labels associated to a value not present in the data.
}
\examples{
x <- labelled(c(1, 2, 2, 1), c(yes = 1, no = 2, maybe = 3))
x
drop_unused_value_labels(x)
}
