% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_na.R
\name{nolabel_to_na}
\alias{nolabel_to_na}
\title{Recode values with no label to NA}
\usage{
nolabel_to_na(x)
}
\arguments{
\item{x}{Object to recode.}
}
\description{
For labelled variables, values with no label will be recoded to \code{NA}.
}
\examples{
v <- labelled(c(1, 2, 9, 1, 9), c(yes = 1, no = 2))
nolabel_to_na(v)
}
