% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R, R/tagged_na.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{labelled}
\alias{is.labelled}
\alias{labelled_spss}
\alias{print_labels}
\alias{\%>\%}
\alias{tagged_na}
\alias{na_tag}
\alias{is_tagged_na}
\alias{format_tagged_na}
\alias{print_tagged_na}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}}

  \item{haven}{\code{\link[haven:format_tagged_na]{format_tagged_na()}}, \code{\link[haven:is.labelled]{is.labelled()}}, \code{\link[haven:is_tagged_na]{is_tagged_na()}}, \code{\link[haven:labelled]{labelled()}}, \code{\link[haven:labelled_spss]{labelled_spss()}}, \code{\link[haven:na_tag]{na_tag()}}, \code{\link[haven:print_labels]{print_labels()}}, \code{\link[haven:print_tagged_na]{print_tagged_na()}}, \code{\link[haven:tagged_na]{tagged_na()}}}
}}

