% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attributes.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Remove attributes}
\usage{
remove_attributes(x, attributes)
}
\arguments{
\item{x}{an object}

\item{attributes}{a character vector indicating attributes to remve}
}
\description{
This function removes specified attributes. When applied to a data.frame, it will also
remove recursively the specified attributes to each column of the data.frame.
}
\examples{
\dontrun{
library(haven)
path <- system.file("examples", "iris.sav", package = "haven")
d <- read_sav(path)
str(d)
d <- remove_attributes(d, "format.spss")
str(d)}
}
