% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_labels_from_df.R
\name{replace_labels_from_df}
\alias{replace_labels_from_df}
\title{Replace Variable Labels in a Data Frame with Labels from A Lookup Data Frame}
\usage{
replace_labels_from_df(
  data,
  label_df,
  label_df_variable_col = "variable",
  label_df_label_col = "label",
  overwrite = TRUE
)
}
\arguments{
\item{data}{data frame for which labels will be replaced}

\item{label_df}{a data frame with a label_df_variable_col and a label_df_label_col}

\item{label_df_variable_col, label_df_label_col}{variable names (as strings) for the lookup of variable names in data, and the corresponding variable label.}

\item{overwrite}{Flag. Whether to remove label if it already exists.}
}
\value{
data, with variable labels replaced
}
\description{
If no replacement exists, keeps the original.
}
