% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_haven_2.0}
\alias{x_haven_2.0}
\alias{x_spss_haven_2.0}
\alias{spss_file}
\alias{dta_file}
\title{Datasets for testing}
\format{
An object of class \code{haven_labelled} of length 6.

An object of class \code{haven_labelled_spss} (inherits from \code{haven_labelled}) of length 10.

An object of class \code{list} of length 13.

An object of class \code{data.frame} with 47 rows and 6 columns.
}
\usage{
x_haven_2.0

x_spss_haven_2.0

spss_file

dta_file
}
\description{
These datasets are used to test compatibility with foreign (spss_foreign),
or haven_2.0 (x_haven_2.0, x_spss_haven_2.0) packages
}
\keyword{datasets}
