% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagged_na.R
\name{unique_tagged_na}
\alias{unique_tagged_na}
\alias{duplicated_tagged_na}
\alias{order_tagged_na}
\alias{sort_tagged_na}
\title{Unique elements, duplicated, ordering and sorting with tagged NAs}
\usage{
unique_tagged_na(x, fromLast = FALSE)

duplicated_tagged_na(x, fromLast = FALSE)

order_tagged_na(
  x,
  na.last = TRUE,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  na_decreasing = decreasing
)

sort_tagged_na(
  x,
  decreasing = FALSE,
  na.last = TRUE,
  na_decreasing = decreasing
)
}
\arguments{
\item{x}{a vector}

\item{fromLast}{logical indicating if duplication should be
considered from the last}

\item{na.last}{if \code{TRUE}, missing values in the data are put last;
if \code{FALSE}, they are put first}

\item{decreasing}{should the sort order be increasing or decreasing?}

\item{method}{the method to be used, see \code{\link[base:order]{base::order()}}}

\item{na_decreasing}{should the sort order for tagged NAs value be
increasing or decreasing?}
}
\description{
These adapations of \code{\link[base:unique]{base::unique()}}, \code{\link[base:duplicated]{base::duplicated()}},
\code{\link[base:order]{base::order()}} and \code{\link[base:sort]{base::sort()}} treats tagged NAs as distinct
values.
}
