% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/val_labels.R
\name{val_labels}
\alias{val_label}
\alias{val_label<-}
\alias{val_labels}
\alias{val_labels<-}
\title{Get / Set value labels}
\usage{
val_labels(x)

val_labels(x) <- value

val_label(x, v)

val_label(x, v) <- value
}
\arguments{
\item{x}{A vector.}

\item{value}{A named vector for \code{val_labels} (see \code{\link{labelled}}) or a character string
for \code{\link{val_labels}}. \code{NULL} to remove the labels.}

\item{v}{A single value.}
}
\value{
\code{val_labels} will return a named vector.

\code{val_label} will return a single character string.
}
\description{
Get / Set value labels
}
\examples{
s <- labelled(c(1, 1, 2), c(Male = 1, Female = 2))
val_labels(s)
}

