% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_na.R
\name{val_labels_to_na}
\alias{val_labels_to_na}
\title{Recode value labels  to NA}
\usage{
val_labels_to_na(x)
}
\arguments{
\item{x}{Object to recode.}

\item{...}{Other arguments passed down to method.}
}
\description{
For labelled variables, values with a label will be recoded to \code{NA}.
}
\examples{
v <- labelled(c(1, 2, 9, 1, 9), c(dk = 9))
val_labels_to_na(v)
}
\seealso{
\code{\link[haven]{zap_labels}}
}
