% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_labels.R
\name{copy_labels}
\alias{copy_labels}
\title{Copy variable and value labels}
\usage{
copy_labels(from, to)
}
\arguments{
\item{from}{Object to copy labels from.}

\item{to}{Object to copu labels to.}
}
\description{
This function copies variable and value labels (including missing values)
from one vector to another or from one data frame to another data frame.
For data frame, labels are copied according to variable names, and only
if variables are the same type in both data frames.
}
\details{
Some base \R functions like \code{\link[base]{subset}} drop variable and
value labels attached to a variable. \code{copy_labels} coud be used
to restore these attributes.
}
