% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to_na.R
\name{missing_to_na}
\alias{missing_to_na}
\title{Recode defined missing values to NA}
\usage{
missing_to_na(x)
}
\arguments{
\item{x}{Object to recode.}
}
\description{
For labelled variables, defined missing values will be recoded to \code{NA}.
}
\examples{
v <- labelled(c(1, 2, 9, 1, 9), c(yes = 1, no = 2, dk = 3), c(F, F, T))
missing_to_na(v)
}

