% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missing_val.R
\name{missing_val}
\alias{missing_val}
\alias{missing_val<-}
\title{Get / Set a self-defined missing value}
\usage{
missing_val(x)

missing_val(x, force = NULL) <- value
}
\arguments{
\item{x}{A vector.}

\item{value}{A vector indicating which values should be considered as missing.}
}
\description{
Get / Set a self-defined missing value
}

