% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{labelled}
\alias{is.labelled}
\alias{labelled_spss}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{haven}{\code{\link[haven]{is.labelled}}, \code{\link[haven]{labelled}}, \code{\link[haven]{labelled_spss}}}
}}

