% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_labels.R
\name{remove_labels}
\alias{remove_labels}
\alias{remove_labels.data.frame}
\alias{remove_labels.default}
\title{Remove variable and value labels}
\usage{
remove_labels(x)

\method{remove_labels}{default}(x)

\method{remove_labels}{data.frame}(x)
}
\arguments{
\item{x}{A vector or a data frame.}
}
\description{
This function removes all variable and values labels.
}

