% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to_labelled.r
\name{to_labelled}
\alias{foreign_to_labelled}
\alias{to_labelled}
\title{Convert variable / value attributes to labelled data}
\usage{
to_labelled(x)

foreign_to_labelled(x)
}
\arguments{
\item{x}{A data set to convert to labelled data frame}
}
\description{
Convert data imported with \code{\link[foreign]{read.spss}} or \code{\link[foreign]{read.dta}}
from \pkg{foreign} package to labelled data, i.e. using \code{\link{labelled}} class.
}
\details{
#' Factors will not be converted. Therefore, you should use \code{use.value.labels = FALSE}
when importing with \code{\link[foreign]{read.spss}} or \code{convert.factors = FALSE} when
importing with \code{\link[foreign]{read.dta}}.

To convert correctly defined missing values imported with \code{\link[foreign]{read.spss}}, you should
have used \code{to.data.frame = FALSE} and \code{use.missings = FALSE}. If you used the option
\code{to.data.frame = TRUE}, meta data describing missing values will not be attached to the import.
If you used \code{use.missings = TRUE}, missing values would have been converted to \code{NA}.

So far, missing values defined in Stata are always imported as \code{NA} by
\code{\link[foreign]{read.dta}} and could not be retrieved by \code{foreign_to_labelled}.
}
\seealso{
\code{\link{labelled}}, \code{\link[foreign]{read.spss}}, \code{\link[foreign]{read.dta}}
}

