% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var_label.R
\name{var_label}
\alias{var_label}
\alias{var_label<-}
\title{Get / Set a variable label}
\usage{
var_label(x)

var_label(x) <- value
}
\arguments{
\item{x}{A vector.}

\item{value}{A character string or \code{NULL} to remove the label.}
}
\description{
Get / Set a variable label
}
\examples{
var_label(iris$Sepal.Length)
var_label(iris$Sepal.Length) <- "Length of the sepal"
\dontrun{
 View(iris)
}
# To remove a variable label
var_label(iris$Sepal.Length) <- NULL
}

