% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_label.R
\name{var_label}
\alias{var_label}
\alias{var_label<-}
\alias{set_variable_labels}
\title{Get / Set a variable label}
\usage{
var_label(x)

var_label(x) <- value

set_variable_labels(.data, ..., .labels = NA)
}
\arguments{
\item{x}{An object.}

\item{value}{A character string or \code{NULL} to remove the label.
For data frames, it could also be a named list or a character vector
of same length as the number of columns in \code{x}.}

\item{.data}{a data frame}

\item{...}{name-value pairs of variable labels (see examples)}

\item{.labels}{variable labels to be applied to the data.frame,
using the same syntax as \code{value} in `var_label(df) <- value`.}
}
\value{
\code{set_variable_labels} will return an updated copy of \code{.data}.
}
\description{
Get / Set a variable label
}
\details{
For data frames, if \code{value} is a named list, only elements whose name will
  match a column of the data frame will be taken into account. If \code{value}
  is a character vector, labels should in the same order as the columns of the
  data.frame.
}
\note{
\code{set_variable_labels} could be used with \code{dplyr}.
}
\examples{
var_label(iris$Sepal.Length)
var_label(iris$Sepal.Length) <- 'Length of the sepal'
\dontrun{
 View(iris)
}
# To remove a variable label
var_label(iris$Sepal.Length) <- NULL
# To change several variable labels at once
var_label(iris) <- c(
  "sepal length", "sepal width", "petal length",
  "petal width", "species"
)
var_label(iris)
var_label(iris) <- list(
  Petal.Width = "width of the petal",
  Petal.Length = "length of the petal"
)
var_label(iris)
if (require(dplyr)) {
  # adding some variable labels
  df <- data_frame(s1 = c("M", "M", "F"), s2 = c(1, 1, 2)) \%>\%
    set_variable_labels(s1 = "Sex", s2 = "Yes or No?")
  var_label(df)

  # removing a variable label
  df <- df \%>\% set_variable_labels(s2 = NULL)
  var_label(df$s2)

  # defining variable labels derived from variable names
  if (require(snakecase)) {
    iris <- iris \%>\%
      set_variable_labels(.labels = to_sentence_case(names(iris)))
    var_label(iris)
  }
}
}
